# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEO Asset Management API",
        "description": "REST API for VEO Asset Management System mobile application",
        "contact": {
            "email": "support@veoset.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://veoset.test",
            "description": "VEO Asset Management API Server"
        }
    ],
    "paths": {
        "/api/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get assets list",
                "operationId": "50c726a8d4ee77df00ce287ce38f08e7",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "description": "Filter by site ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "operational",
                                "maintenance",
                                "emergency",
                                "offline"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assets list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Asset"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get asset details",
                "operationId": "bd8cd9b5700d93a32b257b93822796d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}/status": {
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update asset status",
                "operationId": "3c2d9115d958be589fab5b9f5ee49915",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "operational",
                                            "maintenance",
                                            "emergency",
                                            "offline"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Routine maintenance completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Asset": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Generator Unit 01"
                    },
                    "asset_type": {
                        "type": "string",
                        "example": "generator"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "operational",
                            "maintenance",
                            "emergency",
                            "offline"
                        ],
                        "example": "operational"
                    },
                    "location": {
                        "type": "string",
                        "example": "Building A - Floor 2"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "health_score": {
                        "type": "number",
                        "format": "float",
                        "example": 85.5
                    },
                    "health_status": {
                        "type": "string",
                        "enum": [
                            "excellent",
                            "good",
                            "fair",
                            "poor",
                            "critical"
                        ],
                        "example": "good"
                    },
                    "last_updated": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "site": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Main Facility"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AssetDetail": {
                "properties": {
                    "asset": {
                        "$ref": "#/components/schemas/Asset"
                    },
                    "recent_history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssetHistory"
                        }
                    },
                    "health_trend": {
                        "properties": {
                            "trend": {
                                "type": "string",
                                "enum": [
                                    "improving",
                                    "stable",
                                    "declining"
                                ]
                            },
                            "change": {
                                "type": "number"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AssetHistory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_type": {
                        "type": "string",
                        "enum": [
                            "performance_reading",
                            "diagnostic_scan",
                            "status_change"
                        ],
                        "example": "performance_reading"
                    },
                    "health_score": {
                        "type": "number",
                        "format": "float",
                        "example": 85.5
                    },
                    "performance_score": {
                        "type": "number",
                        "format": "float",
                        "example": 92.1
                    },
                    "temperature": {
                        "type": "number",
                        "format": "float",
                        "example": 45.2
                    },
                    "pressure": {
                        "type": "number",
                        "format": "float",
                        "example": 15.8
                    },
                    "vibration": {
                        "type": "number",
                        "format": "float",
                        "example": 2.1
                    },
                    "anomaly_detected": {
                        "type": "boolean",
                        "example": false
                    },
                    "anomaly_description": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}